refactor(dao): queries * merge query response to msg.rs * move query module to separate file (query.rs) in contract.rs * optimize imports
