+++ contracts/dao/schema/query_msg.json
@@ -3,7 +3,8 @@
   "title": "QueryMsg",
   "oneOf": [
     {
-      "description": "Returns Config",
+      "title": "GetConfig",
+      "description": "Returns [ConfigResponse]\n\n## Example\n\n```json { \"get_config\": {} } ```",
       "type": "object",
       "required": [
         "get_config"
@@ -16,7 +17,8 @@
       "additionalProperties": false
     },
     {
-      "description": "Return list of cw20 Tokens associated with the DAO Treasury",
+      "title": "TokenList",
+      "description": "Returns [TokenListResponse] list of cw20 Tokens associated with the DAO Treasury\n\n## Example\n\n```json { \"token_list\": {} } ```",
       "type": "object",
       "required": [
         "token_list"
@@ -29,7 +31,8 @@
       "additionalProperties": false
     },
     {
-      "description": "Returns All DAO Cw20 Balances",
+      "title": "TokenBalances",
+      "description": "Returns [TokenBalancesResponse] All DAO Cw20 Balances\n\n## Example\n\n```json { \"token_balances\": { \"start\"?: { \"native\": \"uosmo\" | \"cw20\": \"osmo1deadbeef\" }, \"limit\": 30 | 10, \"order\": \"asc\" | \"desc\" } } ```",
       "type": "object",
       "required": [
         "token_balances"
@@ -72,7 +75,8 @@
       "additionalProperties": false
     },
     {
-      "description": "Returns ProposalResponse",
+      "title": "Proposal",
+      "description": "Returns [ProposalResponse]\n\n## Example\n\n```json { \"proposal\": { \"proposal_id\": 1 } } ```",
       "type": "object",
       "required": [
         "proposal"
@@ -95,7 +99,8 @@
       "additionalProperties": false
     },
     {
-      "description": "Returns ProposalListResponse",
+      "title": "Proposals",
+      "description": "Returns [ProposalsResponse]\n\n## Example\n\n```json { \"proposals\": { \"query\": { \"find_by_status\": { \"status\": \"pending\" | .. | \"executed\" } | \"find_by_proposer\": { \"proposer\": \"osmo1deadbeef\" } | \"everything\": {} }, \"start\"?: 10, \"limit\": 30 | 10, \"order\": \"asc\" | \"desc\" } } ```",
       "type": "object",
       "required": [
         "proposals"
@@ -142,7 +147,8 @@
       "additionalProperties": false
     },
     {
-      "description": "Returns the number of proposals in the DAO (u64)",
+      "title": "ProposalCount",
+      "description": "Returns the number of proposals in the DAO (u64)\n\n## Example\n\n```json { \"proposal_count\": {} } ```",
       "type": "object",
       "required": [
         "proposal_count"
@@ -155,7 +161,8 @@
       "additionalProperties": false
     },
     {
-      "description": "Returns VoteResponse",
+      "title": "Vote",
+      "description": "Returns [VoteResponse]\n\n## Example\n\n```json { \"vote\": { \"proposal_id\": 1, \"voter\": \"osmo1deadbeef\" } } ```",
       "type": "object",
       "required": [
         "vote"
@@ -182,7 +189,8 @@
       "additionalProperties": false
     },
     {
-      "description": "Returns VoteListResponse",
+      "title": "Votes",
+      "description": "Returns [VotesResponse]\n\n## Example\n\n```json { \"votes\": { \"proposal_id\": 1, \"start\"?: \"osmo1deadbeef\", \"limit\": 30 | 10, \"order\": \"asc\" | \"desc\" } } ```",
       "type": "object",
       "required": [
         "votes"
@@ -229,6 +237,8 @@
       "additionalProperties": false
     },
     {
+      "title": "Deposit",
+      "description": "Queries single deposit info by proposal id & address of depositor Returns [DepositResponse]\n\n## Example\n\n```json { \"deposit\": { \"proposal_id\": 1, \"depositor\": \"osmo1deadbeef\" } } ```",
       "type": "object",
       "required": [
         "deposit"
@@ -255,6 +265,8 @@
       "additionalProperties": false
     },
     {
+      "title": "Deposits",
+      "description": "Queries multiple deposits info by 1. proposal id 2. depositor address Returns [DepositsResponse]\n\n## Example\n\n```json { \"deposits\": { \"query\": { \"find_by_proposal\": { \"proposal_id\": 1, \"start\"?: \"osmo1deadbeef\" } | \"find_by_depositor\": { \"depositor\": \"osmo1deadbeef\", \"start\"?: \"osmo1deadbeef\" } | \"everything\": { \"start\"?: [1, \"osmo1deadbeef\"] } }, \"limit\": 30 | 10, \"order\": \"asc\" | \"desc\" } } ```",
       "type": "object",
       "required": [
         "deposits"
